+++ haystack/tools/skills/__init__.py
@@ -2,7 +2,8 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-from haystack.tools.skills.skill_store import FileSystemSkillStore, SkillMeta, SkillStore
+from haystack.skill_stores.file_system.skill_store import FileSystemSkillStore
+from haystack.skill_stores.types.protocol import SkillMeta, SkillStore
 from haystack.tools.skills.skill_toolset import SkillToolset
 
 __all__ = ["FileSystemSkillStore", "SkillMeta", "SkillStore", "SkillToolset"]
